Fix  mistake register setting in serial_format()
Currently, there is the issue in register setting into serial_format().
The issue is that parameter for baudrate setting is initialed zero in this function.
In baudrate  is less 9600bps, the issue occurs. In baudrate  is over 9600, not occurs. When call serial_baud() before serial_format(), baudrate override and set with an incorrect value.

Therefore, I fixed register setting not to set the parameter of baudrate setting in serial_format().