Fix up types in SDBlockDevice
Correct types passed to debug() calls - as part of this, block size and
erase size are changed to be 32-bit - using 64-bit variables for these
will cause unnecessary code bloat.

Many uses of bd_size_t in the BlockDevice API should really be uint32_t
or size_t, but that would be a bigger API change.