PSOC6: use 32-bit MCWDT as LP ticker for system deep sleep
lp_ticker in PSoC6 was set to 16 bits, this will overflow in 2 seconds
with a frequency of 32.768Khz, lp_ticker is used for system deep sleep,
so system can sleep for only 2 seconds if 16 bit lp_ticker is used.

Modify lp_ticker to 32 bit, this will let system to be in deep sleep
for around 36 hours, MCWDT Counter2 can be used as a free running
timer to read time and MCWDT Counter0 and Counter 1 can be used to set
up timers for deep-sleep.