STM32WB: Handle re-init case of transport layer
Issue was seen when running BLE_GAP example from
mbed-os-example-ble.

In STM32WB, the M0 core cannot be reset except if the whole target
is reset. So in case of re-initialization of the BLE stack, the
transport layer should not be initialized again. The HCI reset
command will do the job.